# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI CV Evaluation API
  version: 1.0.0
  description: Backend mini project – upload CV & project report, evaluate with RAG + LLM.
servers:
  - url: http://localhost:3000
paths:
...
  /rag/search:
    get:
      summary: Dev utility – semantic search
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Hits
          content:
            application/json:
              schema:
                type: object
                properties:
                  q: { type: string }
                  hits:
                    type: array
                    items:
                      $ref: "#/components/schemas/RagHit"
components:
  schemas:
...
    RagHit:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        raw_text: { type: string }
        distance: { type: number, nullable: true }
...